Add data directory and download script
Rather than mirror the data in the repo, this commit adds a script to
download the outside and inside sniffing tcpdump files from the 1999
DARPA dataset directly from MIT.

Note that it appears a few of the testing archives are actually corrupt,
and we may need to either find mirrors of the valid files, or we may
even need a new dataset. My initial searches didn't find any other
mirrors of the 1999 DARPA dataset though.